fix : 에러 수정
