+++ backend/src/main/java/com/pdfcampus/pdfcampus/repository/NoteRepository.java
@@ -2,6 +2,7 @@ package com.pdfcampus.pdfcampus.repository;
 
 import com.pdfcampus.pdfcampus.entity.Note;
 import org.springframework.data.jpa.repository.JpaRepository;
+import org.springframework.data.jpa.repository.Query;
 import org.springframework.stereotype.Repository;
 
 import java.util.Date;
@@ -16,6 +17,4 @@ public interface NoteRepository extends JpaRepository<Note, Integer> {
 
     List<Note> findByNoteTitleContainingIgnoreCase(String keyword);
     Optional<Note> findByBook(Integer bookId);
-
-    void saveNote(Long userId, Long bookId, String s, String s1, Date date, Date date1);
 }
